Move Bigtable client library into sei-db/db_engine/bigtable
Bundle the Bigtable pieces into a standalone DB library alongside the
other db_engine backends: the gRPC data-plane client (ReadRows chunk
assembly, MutateRows accounting, bounded retries), the frozen row-key
and cell encoding, Config, and the per-RPC cost metrics on the
seidb_bigtable meter. The now-redundant Bigtable prefix is dropped from
names (BigtableClient -> Client, BigtableMutationRowKey ->
MutationRowKey, ...); encoding behavior and metric names are unchanged
and the encoding tests move with the code with assertions untouched.

offload/historical keeps the Reader contract and a slim Bigtable-backed
reader adapter (Get/Has/LastVersion plus mutation-row interpretation);
the offload consumer sink now builds row mutations against the new
package directly.
